EXODUS/IOSS: Fix conversion errors

+++ packages/seacas/libraries/ioss/src/Ioss_Map.C
@@ -1,4 +1,4 @@
-// Copyright(C) 1999-2021 National Technology & Engineering Solutions
+// Copyright(C) 1999-2022 National Technology & Engineering Solutions
 // of Sandia, LLC (NTESS).  Under the terms of Contract DE-NA0003525 with
 // NTESS, the U.S. Government retains certain rights in this software.
 //
@@ -150,8 +150,10 @@ void Ioss::Map::build_reverse_map__(int64_t num_to_get, int64_t offset)
   }
 }
 
-template IOSS_EXPORT bool Ioss::Map::set_map(int *ids, size_t count, size_t offset, bool in_define_mode);
-template IOSS_EXPORT bool Ioss::Map::set_map(int64_t *ids, size_t count, size_t offset, bool in_define_mode);
+template IOSS_EXPORT bool Ioss::Map::set_map(int *ids, size_t count, size_t offset,
+                                             bool in_define_mode);
+template IOSS_EXPORT bool Ioss::Map::set_map(int64_t *ids, size_t count, size_t offset,
+                                             bool in_define_mode);
 
 template <typename INT>
 bool Ioss::Map::set_map(INT *ids, size_t count, size_t offset, bool in_define_mode)
@@ -251,7 +253,7 @@ template <typename INT> void Ioss::Map::reverse_map_data(INT *data, size_t count
   if (!is_sequential()) {
     for (size_t i = 0; i < count; i++) {
       INT global_id = data[i];
-      data[i]       = global_to_local__(global_id, true);
+      data[i]       = (INT)global_to_local__(global_id, true);
     }
   }
   else if (m_offset != 0) {
@@ -283,7 +285,7 @@ template <typename INT> void Ioss::Map::map_data(INT *data, size_t count) const
   IOSS_FUNC_ENTER(m_);
   if (!is_sequential()) {
     for (size_t i = 0; i < count; i++) {
-      data[i] = m_map[data[i]];
+      data[i] = (INT)m_map[data[i]];
     }
   }
   else if (m_offset != 0) {
@@ -306,8 +308,10 @@ void Ioss::Map::map_data(void *data, const Ioss::Field &field, size_t count) con
 }
 
 #ifndef DOXYGEN_SKIP_THIS
-template IOSS_EXPORT void Ioss::Map::map_implicit_data(int *data, size_t count, size_t offset) const;
-template IOSS_EXPORT void Ioss::Map::map_implicit_data(int64_t *data, size_t count, size_t offset) const;
+template IOSS_EXPORT void Ioss::Map::map_implicit_data(int *data, size_t count,
+                                                       size_t offset) const;
+template IOSS_EXPORT void Ioss::Map::map_implicit_data(int64_t *data, size_t count,
+                                                       size_t offset) const;
 #endif
 
 template <typename INT>
@@ -317,12 +321,12 @@ void Ioss::Map::map_implicit_data(INT *ids, size_t count, size_t offset) const
   // ids are implicit
   if (is_sequential()) {
     for (size_t i = 0; i < count; i++) {
-      ids[i] = m_offset + offset + 1 + i;
+      ids[i] = static_cast<INT>(m_offset + offset + 1 + i);
     }
   }
   else {
     for (size_t i = 0; i < count; i++) {
-      ids[i] = m_map[offset + 1 + i];
+      ids[i] = static_cast<INT>(m_map[offset + 1 + i]);
     }
   }
 }
@@ -340,16 +344,20 @@ void Ioss::Map::map_implicit_data(void *data, const Ioss::Field &field, size_t c
 }
 
 template IOSS_EXPORT size_t Ioss::Map::map_field_to_db_scalar_order(double              *variables,
-                                                        std::vector<double> &db_var,
-                                                        size_t begin_offset, size_t count,
-                                                        size_t stride, size_t offset);
-template IOSS_EXPORT size_t Ioss::Map::map_field_to_db_scalar_order(int *variables, std::vector<double> &db_var,
-                                                        size_t begin_offset, size_t count,
-                                                        size_t stride, size_t offset);
+                                                                    std::vector<double> &db_var,
+                                                                    size_t begin_offset,
+                                                                    size_t count, size_t stride,
+                                                                    size_t offset);
+template IOSS_EXPORT size_t Ioss::Map::map_field_to_db_scalar_order(int                 *variables,
+                                                                    std::vector<double> &db_var,
+                                                                    size_t begin_offset,
+                                                                    size_t count, size_t stride,
+                                                                    size_t offset);
 template IOSS_EXPORT size_t Ioss::Map::map_field_to_db_scalar_order(int64_t             *variables,
-                                                        std::vector<double> &db_var,
-                                                        size_t begin_offset, size_t count,
-                                                        size_t stride, size_t offset);
+                                                                    std::vector<double> &db_var,
+                                                                    size_t begin_offset,
+                                                                    size_t count, size_t stride,
+                                                                    size_t offset);
 
 template <typename T>
 size_t Ioss::Map::map_field_to_db_scalar_order(T *variables, std::vector<double> &db_var,
